src/prompt-processing/requirements.txt
pyspellchecker
language_tool_python
langdetect
setuptools